Remove src/components folder and move contents to more suitable locations
- Move SCSS files under `src/components` to `src/scss`
- Move files under CreateSecret to `src/containers/CreateSecret`
- Move files under CreatePipelineResource to `src/containers/CreatePipelineResource`

The reason for moving these and deleting the `src/components` folder is to
help clarify where pure UI components should live (i.e. `packages/components`)
and avoid confusion for contributors.

We may go further in a future change and extract pure UI components from
CreateSecret and/or CreatePipelineResource but that's out of scope for this change.
