+++ .storybook/Container.scss
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2020 The Tekton Authors
+Copyright 2019-2021 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -11,4 +11,4 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-@import '../src/components/App/App';
+@import '../src/scss/App';
