+++ src/containers/ClusterTasks/ClusterTasks.js
@@ -45,7 +45,7 @@ import {
   isWebSocketConnected
 } from '../../reducers';
 
-import '../../components/Definitions/Definitions.scss';
+import '../../scss/Definitions.scss';
 
 export /* istanbul ignore next */ class ClusterTasksContainer extends Component {
   state = {
