+++ src/containers/CreatePipelineResource/CreatePipelineResource.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2020 The Tekton Authors
+Copyright 2019-2021 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -22,8 +22,8 @@ import {
   urls
 } from '@tektoncd/dashboard-utils';
 
-import GitResourceFields from '../../components/CreatePipelineResource/GitResourceFields';
-import UniversalFields from '../../components/CreatePipelineResource/UniversalFields';
+import GitResourceFields from './GitResourceFields';
+import UniversalFields from './UniversalFields';
 import { createPipelineResource } from '../../api';
 import { getSelectedNamespace } from '../../reducers';
