+++ src/containers/CreateSecret/Form/UniversalFields.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019-2020 The Tekton Authors
+Copyright 2019-2021 The Tekton Authors
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
@@ -20,7 +20,7 @@ import {
   TextInput,
   Tooltip
 } from 'carbon-components-react';
-import NamespacesDropdown from '../../../containers/NamespacesDropdown';
+import NamespacesDropdown from '../../NamespacesDropdown';
 
 const UniversalFields = props => {
   const {
