+++ src/containers/Tasks/Tasks.js
@@ -47,7 +47,7 @@ import {
   isWebSocketConnected
 } from '../../reducers';
 
-import '../../components/Definitions/Definitions.scss';
+import '../../scss/Definitions.scss';
 
 export /* istanbul ignore next */ class Tasks extends Component {
   state = {
